craw-data/requirements.txt
# ============================================================
# NLU-EduPath – Crawl Data
# Python 3.12+
# ============================================================

# --- Scraping Core ---
Scrapy==2.11.2
scrapy-playwright==0.0.40
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0
itemadapter==0.9.0
w3lib==2.2.1

# --- HTTP ---
httpx==0.28.1
requests==2.32.3
fake-useragent==2.0.3
scrapy-rotating-proxies==0.6.2

# --- Database ---
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# --- Validation ---
pydantic==2.10.4
pydantic-settings==2.7.0

# --- Task Queue & Scheduling ---
celery==5.4.0
redis==5.2.1
APScheduler==3.11.0

# --- Config & Environment ---
python-dotenv==1.0.1

# --- Logging ---
colorlog==6.9.0

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
responses==0.25.3

# --- Dev Tools ---
ipython==8.30.0
